- refactored code due to AI review

+++ src/main/resources/init_db.sql
@@ -1,5 +1,5 @@
 CREATE TABLE books(
     id BIGINT AUTO_INCREMENT PRIMARY KEY,
     title varchar(255),
-    price float
+    price DECIMAL(10,2)
 )
